requirements/base.txt
redis==4.5.4  # https://github.com/redis/redis-py
hiredis==2.2.2  # https://github.com/redis/hiredis-py
celery==5.2.7
pytz==2023.3
flower==1.2.0
phonenumbers==8.13.10
Pillow==9.5.0
whitenoise==6.4.0  # https://github.com/evansd/whitenoise
hiredis==2.2.2  # https://github.com/redis/hiredis-py

# DJANGO    
#--------------------------------------------------------------
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
Django==4.2
django-celery-beat==2.5.0
django-cleanup==7.0.0
django-countries==7.5.1
django-filter==23.1
django-model-utils==4.3.1
django-money==3.1.0
django-phonenumber-field==7.1.0
django-redis==5.2.0
django-storages==1.13.2

# Django REST Framework
djangorestframework==3.14.0
djangochannelsrestframework==1.1.0
django-cors-headers==3.14.0

# Authentication
djoser==2.2.0
djangorestframework-simplejwt==5.2.2
django-channels-jwt-auth-middleware==1.0.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2

# DRF-spectacular for api documentation
drf-spectacular==0.26.2
drf-spectacular-sidecar==2023.4.1